requirements.txt
Flask
gunicorn
google-genai
PIL
climage